assign_4/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
opencv-python>=4.5.0
pillow>=8.0.0